Drop prerelease version suffix.
Six months should be long enough to be confident this release is safe, don't you think? :)

This came up because of a bug involving `npm`, `semver`, `read-installed`, and prerelease packages: https://github.com/npm/read-installed/issues/40

If you decide to accept this commit, please publish a new release to NPM as soon as possible. Thanks!